# swagger.yaml
openapi: 3.0.3
info:
  title: Elasticsearch PoC API
  description: |
    A comprehensive Proof of Concept API for evaluating Elasticsearch as a search and analytics engine.
    
    This API provides endpoints for:
    - Product management (CRUD operations)
    - Advanced search capabilities with multiple query types
    - Analytics and aggregations
    - Performance testing and monitoring
    
    ## Features Tested
    - Full-text search with relevance scoring
    - Multi-field search across name, description, category
    - Fuzzy search with auto-correction
    - Wildcard and regex pattern matching
    - Boolean queries with must/should/filter clauses
    - Range filters for price, rating, dates
    - Faceted search with aggregation-based filters
    - Statistical and bucket aggregations
    - Performance benchmarking and load testing
    
  version: 1.0.0
  contact:
    name: Lawsikho Team
    email: support@lawsikho.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.example.com
    description: Production server (example)
...
paths:
...
  /api/products:
    get:
      tags:
        - Products
      summary: Get all products
      description: Retrieve products with pagination and sorting
      parameters:
        - name: page
          in: query
          description: Page number (1-based)
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: size
          in: query
          description: Number of products per page
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: sort
          in: query
          description: Field to sort by
          required: false
          schema:
            type: string
            enum: [name, price, rating, createdAt, updatedAt]
            default: createdAt
        - name: order
          in: query
          description: Sort order
          required: false
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: Products retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  products:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'
                  total:
                    type: integer
                    example: 1000
                  page:
                    type: integer
                    example: 1
                  size:
                    type: integer
                    example: 10
                  totalPages:
                    type: integer
                    example: 100
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags:
        - Products
      summary: Create new product
      description: Create a new product in the system
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductInput'
      responses:
        '201':
          description: Product created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/products/{id}:
    get:
      tags:
        - Products
      summary: Get product by ID
      description: Retrieve a specific product by its ID
      parameters:
        - name: id
          in: path
          required: true
          description: Product ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Product found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags:
        - Products
      summary: Update product
      description: Update an existing product
      parameters:
        - name: id
          in: path
          required: true
          description: Product ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductInput'
      responses:
        '200':
          description: Product updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Products
      summary: Delete product
      description: Delete a product from the system
      parameters:
        - name: id
          in: path
          required: true
          description: Product ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Product deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Product deleted successfully"
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/products/bulk:
    post:
      tags:
        - Products
      summary: Bulk create products
      description: Create multiple products in a single operation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                products:
                  type: array
                  items:
                    $ref: '#/components/schemas/ProductInput'
                  maxItems: 1000
              required:
                - products
      responses:
        '200':
          description: Bulk operation completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Bulk insert completed"
                  total:
                    type: integer
                    example: 100
                  errors:
                    type: boolean
                    example: false
                  items:
                    type: integer
                    example: 100
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/products/stats:
    get:
      tags:
        - Products
      summary: Get index statistics
      description: Retrieve Elasticsearch index statistics
      responses:
        '200':
          description: Statistics retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_documents:
                    type: integer
                    example: 1000
                  index_size:
                    type: integer
                    example: 1048576
                  shards:
                    type: object
                    properties:
                      total:
                        type: integer
                        example: 1
                      successful:
                        type: integer
                        example: 1
                      failed:
                        type: integer
                        example: 0
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/performance/load-test:
    post:
      tags:
        - Performance
      summary: Run load test
      description: Run custom load tests with specified queries
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoadTestRequest'
      responses:
        '200':
          description: Load test completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoadTestResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/performance/benchmark:
    get:
      tags:
        - Performance
      summary: Run benchmark tests
      description: Run standard performance benchmark tests
      parameters:
        - name: size
          in: query
          description: Dataset size for benchmarking
          schema:
            type: integer
            minimum: 100
            maximum: 10000
            default: 1000
      responses:
        '200':
          description: Benchmark completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BenchmarkResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/performance/index-stats:
    get:
      tags:
        - Performance
      summary: Get index statistics
      description: Get detailed Elasticsearch index performance statistics
      responses:
        '200':
          description: Index statistics retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IndexStats'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
    Product:
      type: object
      required:
        - id
        - name
        - category
        - price
        - rating
        - inStock
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
          description: Unique product identifier
          example: "123e4567-e89b-12d3-a456-426614174000"
        name:
          type: string
          description: Product name
          example: "Smart Wireless Headphones"
        description:
          type: string
          description: Product description
          example: "High-quality wireless headphones with noise cancellation"
        category:
          type: string
          description: Product category
          example: "Electronics"
        price:
          type: number
          format: double
          description: Product price
          example: 299.99
        rating:
          type: number
          format: double
          minimum: 0
          maximum: 5
          description: Product rating
          example: 4.5
        tags:
          type: array
          items:
            type: string
          description: Product tags
          example: ["wireless", "premium", "noise-cancellation"]
        inStock:
          type: boolean
          description: Product availability
          example: true
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
          example: "2024-01-01T00:00:00.000Z"
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
          example: "2024-01-01T00:00:00.000Z"
        metadata:
          type: object
          properties:
            brand:
              type: string
              example: "TechBrand"
            color:
              type: string
              example: "Black"
            size:
              type: string
              example: "M"
            weight:
              type: number
              format: double
              example: 1.5
        _score:
          type: number
          description: Search relevance score
          example: 0.95
    ProductInput:
      type: object
      required:
        - name
        - category
        - price
        - rating
        - inStock
      properties:
        name:
          type: string
          description: Product name
          example: "Smart Wireless Headphones"
        description:
          type: string
          description: Product description
          example: "High-quality wireless headphones with noise cancellation"
        category:
          type: string
          description: Product category
          example: "Electronics"
        price:
          type: number
          format: double
          description: Product price
          example: 299.99
        rating:
          type: number
          format: double
          minimum: 0
          maximum: 5
          description: Product rating
          example: 4.5
        tags:
          type: array
          items:
            type: string
          description: Product tags
          example: ["wireless", "premium", "noise-cancellation"]
        inStock:
          type: boolean
          description: Product availability
          example: true
        metadata:
          type: object
          properties:
            brand:
              type: string
              example: "TechBrand"
            color:
              type: string
              example: "Black"
            size:
              type: string
              example: "M"
            weight:
              type: number
              format: double
              example: 1.5
...
    LoadTestRequest:
      type: object
      required:
        - queries
      properties:
        queries:
          type: array
          items:
            type: object
            description: Elasticsearch query object
          minItems: 1
          maxItems: 10
        iterations:
          type: integer
          minimum: 1
          maximum: 100
          default: 10
          description: Number of test iterations
        concurrent:
          type: integer
          minimum: 1
          maximum: 50
          default: 5
          description: Number of concurrent requests
        testType:
          type: string
          enum: [search, index, get]
          default: search
          description: Type of test to run
    LoadTestResponse:
      type: object
      properties:
        testType:
          type: string
          example: "search"
        iterations:
          type: integer
          example: 10
        concurrent:
          type: integer
          example: 5
        totalQueries:
          type: integer
          example: 5
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        statistics:
          type: object
          properties:
            totalQueries:
              type: integer
              example: 50
            successfulQueries:
              type: integer
              example: 48
            failedQueries:
              type: integer
              example: 2
            successRate:
              type: number
              format: double
              example: 96.0
            averageDuration:
              type: number
              format: double
              example: 25.5
            minDuration:
              type: integer
              example: 15
            maxDuration:
              type: integer
              example: 45
            p50Duration:
              type: integer
              example: 23
            p90Duration:
              type: integer
              example: 35
            p95Duration:
              type: integer
              example: 40
            p99Duration:
              type: integer
              example: 44
        results:
          type: array
          items:
            type: object
            properties:
              iteration:
                type: integer
                example: 1
              queries:
                type: array
                items:
                  type: object
                  properties:
                    query:
                      type: integer
                      example: 1
                    success:
                      type: boolean
                      example: true
                    duration:
                      type: integer
                      example: 25
                    responseSize:
                      type: integer
                      example: 1024
                    hits:
                      type: integer
                      example: 10
                    took:
                      type: integer
                      example: 20
    BenchmarkResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        tests:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
                example: "Simple Match Query"
              success:
                type: boolean
                example: true
              duration:
                type: integer
                example: 25
              elasticsearchTook:
                type: integer
                example: 20
              totalHits:
                type: integer
                example: 150
              responseSize:
                type: integer
                example: 1024
        summary:
          type: object
          properties:
            totalTests:
              type: integer
              example: 6
            successfulTests:
              type: integer
              example: 6
            failedTests:
              type: integer
              example: 0
            averageDuration:
              type: number
              format: double
              example: 28.5
            minDuration:
              type: integer
              example: 15
            maxDuration:
              type: integer
              example: 45
    IndexStats:
      type: object
      properties:
        index:
          type: object
          properties:
            name:
              type: string
              example: "products"
            total_documents:
              type: integer
              example: 1000
            deleted_documents:
              type: integer
              example: 0
            index_size_bytes:
              type: integer
              example: 1048576
            index_size_mb:
              type: number
              format: double
              example: 1.0
        shards:
          type: object
          properties:
            total:
              type: integer
              example: 1
            successful:
              type: integer
              example: 1
            failed:
              type: integer
              example: 0
        primaries:
          type: object
          properties:
            documents:
              type: integer
              example: 1000
            size_bytes:
              type: integer
              example: 1048576
            size_mb:
              type: number
              format: double
              example: 1.0
        search:
          type: object
          properties:
            query_total:
              type: integer
              example: 500
            query_time_ms:
              type: integer
              example: 1250
            fetch_total:
              type: integer
              example: 500
            fetch_time_ms:
              type: integer
              example: 750
        indexing:
          type: object
          properties:
            index_total:
              type: integer
              example: 1000
            index_time_ms:
              type: integer
              example: 2000
            index_current:
              type: integer
              example: 0
    Error:
      type: object
      required:
        - error
        - timestamp
      properties:
        error:
          type: string
          description: Error message
          example: "Internal Server Error"
        message:
          type: string
          description: Detailed error message
          example: "Something went wrong"
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
          example: "2024-01-01T00:00:00.000Z"
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication
security:
  - ApiKeyAuth: []